Created edit tests for scenario controller
These tests are not very rigorous.
Not all possiblities are or paths in the controllers are tested for.
The best way to do this would be to refactor our code into services and repositories.
Then, test the services and controllers.
The tests would be much smaller and could be made to test the logic and interconnections easily.
Right now, our controllers act kind of like a mini god class for their respective models and views.